hams-stock-worker/requirements.txt
fastapi
uvicorn
yt-dlp
openai-whisper
torch
firebase-admin 
httpx 
feedparser 
python-dotenv